monetization/backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<3.14
psycopg2-binary>=2.9,<3.0 # Для PostgreSQL, если используется
decouple>=0.5 # Для управления настройками через переменные окружения
# requests>=2.25 # Для взаимодействия с внешними API (например, платежными шлюзами)
# celery>=5.0 # Для фоновых задач, если потребуется
# redis>=3.5 # Для Celery, если используется
# gunicorn>=20.1 # Для развертывания